Feat [#6] collectionview에 header 추가
